Changed: buffer blocking policy adjusted for update write
keys:
1、When the update write (pre read) is not complete, block the buffer;
2、buffer blocking policy
(1) first check the buffer is full, if dissatisfied, check whether the
current request to put down the data, if so, put the current request, if
not, then block the buffer;

(2) If buffer is blocked, select the replacement of the two ends of the
page. If the two full page, then issued together to lift the buffer
block; if a partial page 1 full page or 2 partial page, then issued a
pre-read request, waiting for the completion of full page and then
issued , And then release the buffer block.
3、Trace read is complete, the request queue is still incomplete request,
this time still to implement buff
Bug fixed:To solve the last trace does not perform
Sloved:the reason to ensure that the last trace time is greater than all
the previous state transition time
Problem:When testing large trace files, there is a problem with memory
leaks ep:f2 trace
16M_2KB_sequence_RandW/update_16_count/example.ascii run successful